% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_data_summary}
\alias{get_data_summary}
\title{Get Data Summary}
\source{
Get Summary Data pulls data from the following reports:
\href{https://data.ca.gov/dataset/drinking-water-public-water-system-operations-monthly-water-production-and-conservation-information}{Conservation Report (CR)}
\href{https://www.waterboards.ca.gov/drinking_water/certlic/drinkingwater/eardata.html}{Electronic Annual Report (EAR)}
\href{https://wuedata.water.ca.gov/}{Urban Water Management Plan (UWMP)}
\href{https://wuedata.water.ca.gov/public/awwa_data_export/water_audit_data_conv_to_af.xls}{Water Loss Audit (WLA)}
}
\usage{
get_data_summary(
  category_selection = c("supply", "demand", "supply total", "demand total", "losses",
    "other"),
  report_selection = c("EAR", "UWMP", "CR", "WLA")
)
}
\arguments{
\item{category_selection}{Describes type of data to pull. an be \code{"supply"},
\code{"supply total"}, \code{"demand"}, \code{"demand total"}, \code{"losses"}, \code{other}}

\item{report_selection}{Specifies which report to pull data from. Can be one of \code{"EAR"}, \code{"UWMP"}
\code{"CR"}, or \code{"WLA"}}
}
\description{
Summarizes data from the CR, EAR, UWMP, and WLA into summary table.
}
