% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_ear_data}
\alias{get_ear_data}
\title{Get ear data}
\source{
# TODO add source
}
\usage{
get_ear_data(year_selection)
}
\arguments{
\item{year_selection}{Year to filter data to.}
}
\description{
This function pulls ear data This is a helper function used by the \code{pull_data()} function
}
