% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_wla_data}
\alias{get_wla_data}
\title{Get wla data}
\source{
\href{https://wuedata.water.ca.gov/public/awwa_data_export/water_audit_data_conv_to_af.xls}{Water Loss Audit (WLA)}
}
\usage{
get_wla_data()
}
\description{
This function pulls wla data from wue portal. This is a helper function used by the \code{pull_data()} function
}
