% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_county_data}
\alias{map_county_data}
\title{Map Urban Water Data}
\usage{
map_county_data(data)
}
\arguments{
\item{data}{Data to plot. Must be in the format outputted from \code{pull_data()} function}
}
\description{
This function generates a county-level map of Urban Water Data.
}
\examples{
map_data(data)
}
