% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Urban Water Data}
\usage{
plot_data(
  selected_category = c("supply", "supply total", "demand", "demand total", "losses"),
  pwsid
)
}
\arguments{
\item{selected_category}{The category that the user would like plotted.
Coverage accross reporting requirements varies by category.
Can be \code{"supply"}, \code{"supply total"}, \code{"demand"}, \code{"demand total"}, \code{"losses"}}

\item{pwsid}{The Public Water Sysetm Identification Number (PWSID) corresponding
to the agency whose data you would like to visualize.}

\item{mode}{The mode to run model in. Can be \code{"seed"}, \code{"simulate"}, \code{"calibrate"}}
}
\description{
This function generates a simple visualization of Urban Water Data.
}
\examples{
plot_data(selected_category = "demand", pwsid = "CA3610001")
}
