% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Data}
\source{
TODO add
}
\usage{
pull_data(
  type = c("supply", "demand", "supply total", "demand total", "losses", "other"),
  year_selection,
  report = c("EAR", "UWMP", "CR", "WLA"),
  pwsid,
  ...
)
}
\arguments{
\item{type}{Describes type of data to pull. an be \code{"supply"},
\code{"supply total"}, \code{"demand"}, \code{"demand total"}, \code{"losses"}, \code{other}}

\item{year_selection}{Year to filter data to.}

\item{report}{Specifies which report to pull data from. Can be one of \code{"EAR"}, \code{"UWMP"}
\code{"CR"}, or \code{"WLA"}}
}
\description{
Pulls data from the CR, EAR, UWMP, and WLA and compiles it into a single format.
}
