% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwsid_lookup}
\alias{pwsid_lookup}
\title{PWSID Lookup Table}
\format{
A table with 2 columns (\code{water_system_name} & \code{pwsid}) and 7,898 rows
}
\source{
\href{https://data.ca.gov/dataset/drinking-water-public-water-system-information/resource/9dca2f92-4630-4bee-a9f9-69d2085b57e3}{CA open data portal}
}
\usage{
pwsid_lookup
}
\description{
This lookup table helps a user find a PWSID for a specific agency using the agency's name.
}
\keyword{datasets}
