% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{use_type_lookup}
\alias{use_type_lookup}
\title{Use Type Lookup}
\format{
A tibble with 4 columns (\code{use_group}, \code{report_name}, \code{use_type}, \code{category}) and 141 rows
}
\usage{
use_type_lookup
}
\description{
This lookup table helps map use types between reports
}
\keyword{datasets}
