% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_cr_data}
\alias{get_cr_data}
\title{Get CR data}
\source{
# TODO add source
}
\usage{
get_cr_data()
}
\arguments{
\item{year_selection}{Year to filter data to.}
}
\description{
This function pulls cr data from data.ca.gov. This is a helper function used by the \code{pull_data()} function
}
