% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_uwmp_data}
\alias{get_uwmp_data}
\title{Get uwmp data}
\source{
WUE data portal https://wuedata.water.ca.gov/
UWMP required every five years
2020 data comes from this site https://data.cnra.ca.gov/dataset/2020-uwmp-data-export-tables
2015 data comes from this site https://wuedata.water.ca.gov/uwmp_export.asp
\href{https://wuedata.water.ca.gov/}{Urban Water Management Plan (UWMP)}
}
\usage{
get_uwmp_data()
}
\description{
This function pulls uwmp data from wue portal. This is a helper function used by the \code{pull_data()} function
}
