% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{get_wla_data}
\alias{get_wla_data}
\title{Get wla data}
\source{
# TODO add source
}
\usage{
get_wla_data()
}
\arguments{
\item{year_selection}{Year to filter data to.}
}
\description{
This function pulls wla data from wue portal. This is a helper function used by the \code{pull_data()} function
}
