% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Parameters}
\format{
A tibble with 4 columns (\code{year}, \code{url}, \code{file_name}, \code{report_name}) and 23 rows
}
\usage{
parameters
}
\description{
This table provides links to all the web pages used to download urban water data.
}
\keyword{datasets}
