% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Urban Water Data}
\usage{
plot_data(
  data,
  category_selection = c("supply", "supply total", "demand", "demand total", "losses"),
  pwsid_selection,
  show_subcategories = FALSE
)
}
\arguments{
\item{data}{Data to plot. Must be in the format outputted from \code{pull_data()} function}

\item{category_selection}{The category that the user would like plotted.
Coverage across reporting requirements varies by category.
Can be \code{"supply"}, \code{"supply total"}, \code{"demand"}, \code{"demand total"}, \code{"losses"}}

\item{pwsid_selection}{The Public Water System Identification Number (PWSID) corresponding
to the agency whose data you would like to visualize.}

\item{show_subcategories}{Defaults to FALSE. If set to TRUE shows subcategories for use types.}
}
\description{
This function generates a simple visualization of Urban Water Data.
}
\examples{
\dontrun{
pull_data(2020) |>
    plot_data(category_selection = "demand", pwsid = "CA3610001")
}
}
