% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Data}
\source{
Get Summary Data pulls data from the following reports:
\href{https://data.ca.gov/dataset/drinking-water-public-water-system-operations-monthly-water-production-and-conservation-information}{Conservation Report (CR)},
\href{https://www.waterboards.ca.gov/drinking_water/certlic/drinkingwater/eardata.html}{Electronic Annual Report (EAR)},
\href{https://wuedata.water.ca.gov/}{Urban Water Management Plan (UWMP)},
\href{https://wuedata.water.ca.gov/public/awwa_data_export/water_audit_data_conv_to_af.xls}{Water Loss Audit (WLA)}
}
\usage{
pull_data(
  category_selection = c("supply", "demand", "supply total", "demand total", "losses",
    "other"),
  year_selection,
  report_selection = c("EAR", "UWMP", "CR", "WLA"),
  pwsid,
  ...
)
}
\arguments{
\item{category_selection}{Describes type of data to pull. an be \code{"supply"},
\code{"supply total"}, \code{"demand"}, \code{"demand total"}, \code{"losses"}, \code{other}}

\item{year_selection}{One year to filter data to (ex \code{2016}).}

\item{report_selection}{Specifies which report to pull data from. Can be one of \code{"EAR"}, \code{"UWMP"}
\code{"CR"}, or \code{"WLA"}}
}
\value{
A table containing 8 columns: \code{report_name}, \code{pwsid}, \code{supplier_name}, \code{year}, \code{month},
\code{category}, \code{volume_af}, \code{use_group}
}
\description{
Pull one year of data from the CR, EAR, UWMP, and WLA and compiles it into a single format.
}
